Check if the supplied date is valid

+++ suspend.sh
@@ -3,20 +3,26 @@ export sdate=$1
 
 if [[ -z "$@" ]]; then
     echo >&2 "You must supply an argument!"
-    echo >&2 "Usage suspend.sh \"time in words\" see php date"
+    echo >&2 "Usage suspend.sh \"time(see php date)\""
     exit 1
 fi
-
-echo "---MARK: `date +%d-%m-%Y`---"
-
-# Get root
-sudo killall -qw rtcwake
 
 date=`/usr/bin/php << 'EOF'
 <?php
 $date = strtotime(GETENV("sdate"));
 echo "\r".$date;
 EOF`
+
+# Check if a valid date has been supplied
+if ! [[ "$date" = *[[:digit:]]* ]]; then
+    echo >&2 "Invalid date supplied"
+    exit 1
+fi
+
+# Get root
+sudo killall -qw rtcwake
+
+echo "---MARK: `date +%d-%m-%Y`---"
 
 # Set the alarm to wake-up the system
 echo "[`date +%H:%M:%S`] `sudo rtcwake -d rtc0 -m no -t $date`"
